requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.6
sqlparse==0.5.3
tzdata==2024.2
